pyproject.toml
[tool.poetry]
name = "agentictrust"
version = "0.1.0"
description = "Secure OAuth Framework for LLM-Based Agents"
authors = ["Subramanya N <subramanyanagabhushan@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
pyjwt = "2.8.0"
requests = "2.31.0"
python-dotenv = "1.0.0"
werkzeug = "2.3.7"
loguru = "^0.7.2"
pyyaml = "^6.0"
httpx = "^0.28.1"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
sqlalchemy = "^2.0.40"
pydantic = "^2.11.3"
cryptography = "^44.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88